refactor(example): restructure secret management and configuration files for improved organization
