+++ examples/with-secret-registry/src/compose-stacks.ts
@@ -1,7 +1,7 @@
 import { NamespaceStack } from '@kubricate/stacks';
 import { AppStack } from './stacks/AppStack';
-import { config } from './shared';
-import { secretRegistry } from './secret';
+import { config } from './shared-config';
+import { secretRegistry } from './setup-secret';
 
 const namespace = new NamespaceStack().from({
   name: config.namespace,
