+++ examples/with-secret-registry/src/setup-secret.ts
