+++ examples/with-secret-registry/src/shared-config.ts
