+++ examples/with-secret/src/compose-stacks.ts
@@ -1,7 +1,8 @@
 import { NamespaceStack } from '@kubricate/stacks';
 import { AppStack } from './stacks/AppStack';
-import { config, secretManager } from './config';
+import { secretManager } from './setup-secrets';
 import { CronJobStack } from './stacks/CronJobStack';
+import { config } from './shared-config';
 
 const namespace = new NamespaceStack().from({
   name: config.namespace,
