+++ examples/with-secret/src/shared-config.ts
@@ -1,0 +1,3 @@
+export const config = {
+  namespace: 'my-namespace',
+};
